refactor(freezer): move BulkResultText out of presentation, log the silent catch
BulkResultNotifier lives in data/ and imported
com.valhalla.thor.presentation.tile.bulkResultMessage — the only data ->
presentation import anywhere under data/. This project ran a whole PR on layer
purity, so that is a convention regression introduced by this branch, not a
pre-existing wart.

bulkResultMessage is rendered by both the tile and the notifier, so it belongs
with UiText, which it already returns. Moved to com.valhalla.thor.util alongside
it, with the test. No behaviour change.

Also logs BulkResultNotifier's empty catch (SecurityException) — every other
catch on this branch logs, and a silently swallowed revocation race is exactly
the kind of thing worth seeing in a bug report.
